B9IS109 WEB DEVELOPMENT FOR INFORMATION SYSTEMS CA1
RMA Project
RMA (Return Merchandise Authorization) is the process of specifying the product or products purchased by the customer as defective and maintaining and returning them.
Within the scope of this project, the focus was on the products of a company that provides support in the field of computers and computer networks. An attempt was made to create a more secure and user-friendly Return Merchandise Authorization (RMA) tracking system, replacing traditional Excel spreadsheets.

MUSTAFA SEVER 120018641